fixed PI; PI, EULER, gamma, abs were tested, implementing polar

+++ MathParser/CUDA_Math_Parser/cuda_tokens.cuh
@@ -260,7 +260,7 @@ public:
 		return 0;
 	}
 private:
-	T m_val = std::acos(-1);
+	T m_val = acosf(-1);
 };
 
 template <class T>
